Add support for controller functions with prototypes.
This change adds support for controller functions that have functions
defined in their prototype. This works correctly in angular by default because
the __proto__ property is set correctly when angular loads the controller.
When the controller is loaded by requirejs the __proto__ property is set to
the prototype of the intermediate controller function not the function loaded by requirejs.